Update all Associate layers

+++ src/main/java/io/droksty/transfersdemo/service/AssociateServiceImpl.java
@@ -4,6 +4,7 @@ import io.droksty.transfersdemo.dto.AssociateDTO;
 import io.droksty.transfersdemo.model.Associate;
 import io.droksty.transfersdemo.repository.AssociateRepository;
 import io.droksty.transfersdemo.util.Mapper;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,8 +12,6 @@ import java.util.List;
 
 @Service
 public class AssociateServiceImpl implements AssociateService {
-
-    // Section
     private final AssociateRepository associateRepository;
 
     @Autowired
@@ -20,22 +19,28 @@ public class AssociateServiceImpl implements AssociateService {
         this.associateRepository = associateRepository;
     }
 
+    /* -- Method Implementations -- */
 
-    // Method Implementations
     @Override
     public Associate insertAssociate(AssociateDTO associateDTO) {
+        if (associateDTO.getId() != null)
+            throw new IllegalArgumentException("Associate id is already set");
         Associate associate = Mapper.newAssociateFrom(associateDTO);
         return associateRepository.save(associate);
     }
 
     @Override
     public Associate updateAssociate(AssociateDTO associateDTO) {
+        if (!existsById(associateDTO.getId()))
+            throw new EntityNotFoundException("Associate with id " + associateDTO.getId() + " not found");
         Associate associate = Mapper.newAssociateFrom(associateDTO);
         return associateRepository.save(associate);
     }
 
     @Override
     public void deleteAssociate(Long id) {
+        if (!existsById(id))
+            throw new EntityNotFoundException("Associate with id " + id + " was not found.");
         associateRepository.deleteById(id);
     }
 
@@ -48,5 +53,4 @@ public class AssociateServiceImpl implements AssociateService {
     public List<Associate> getAllAssociates() {
         return associateRepository.findAll();
     }
-
 }
